refactor: centralize sanitizeForFirestore and guard stripUndefinedDeep against non-plain objects
